requirements.txt
nose
zope.component[zcml]
psycopg2

